+++ packages/vite/index.cjs
@@ -1,6 +1,12 @@
 /* eslint-disable no-restricted-globals */
-// Proxy ESM to be used in CJS
 
+// proxy cjs utils
+Object.assign(module.exports, require('./dist/node-cjs/utils.cjs'))
+
+// type utils
+module.exports.defineConfig = (config) => config
+
+// async functions, can be redirect from ESM build
 module.exports.build = (...args) =>
   import('./dist/node/index.js').then((i) => i.build(...args))
 
@@ -13,24 +19,21 @@ module.exports.preview = (...args) =>
 module.exports.transformWithEsbuild = (...args) =>
   import('./dist/node/index.js').then((i) => i.transformWithEsbuild(...args))
 
+module.exports.resolveConfig = (...args) =>
+  import('./dist/node/index.js').then((i) => i.resolveConfig(...args))
+
+module.exports.optimizeDeps = (...args) =>
+  import('./dist/node/index.js').then((i) => i.optimizeDeps(...args))
+
 module.exports.formatPostcssSourceMap = (...args) =>
   import('./dist/node/index.js').then((i) => i.formatPostcssSourceMap(...args))
 
-module.exports.splitVendorChunkPlugin = (...args) =>
-  require('./dist/node-cjs/splitVendorChunk.cjs').splitVendorChunkPlugin(
-    ...args
-  )
-
-module.exports.defineConfig = (config) => config
-
-const os = require('os')
-const path = require('path')
-const isWindows = os.platform() === 'win32'
-
-function slash(p) {
-  return p.replace(/\\/g, '/')
-}
-
-module.exports.normalizePath = (id) => {
-  return path.posix.normalize(isWindows ? slash(id) : id)
-}
+// some sync functions are marked not supported due to their complexity and uncommon usage
+const unsupportedCJS = ['resolvePackageEntry', 'resolvePackageData']
+unsupportedCJS.forEach((name) => {
+  module.exports[name] = (...args) => {
+    throw new Error(`"${name}" is not supported in CJS build of Vite 3.
+Please use ESM or dynamic imports \`const { ${name} } = await import('vite')\`.
+`)
+  }
+})
