+++ packages/vite/src/node/config.ts
@@ -26,6 +26,7 @@ import {
   isExternalUrl,
   isObject,
   lookupFile,
+  mergeAlias,
   normalizePath
 } from './utils'
 import { resolvePlugins } from './plugins'
@@ -40,6 +41,7 @@ import type { JsonOptions } from './plugins/json'
 import type { PluginContainer } from './server/pluginContainer'
 import { createPluginContainer } from './server/pluginContainer'
 import type { PackageCache } from './packages'
+import { mergeConfig, normalizeAlias } from './utils'
 import type { ResolvedBuildOptions } from '.'
 
 const debug = createDebugger('vite:config')
@@ -617,117 +619,6 @@ function resolveBaseUrl(
   return base
 }
 
-function mergeConfigRecursively(
-  defaults: Record<string, any>,
-  overrides: Record<string, any>,
-  rootPath: string
-) {
-  const merged: Record<string, any> = { ...defaults }
-  for (const key in overrides) {
-    const value = overrides[key]
-    if (value == null) {
-      continue
-    }
-
-    const existing = merged[key]
-
-    if (existing == null) {
-      merged[key] = value
-      continue
-    }
-
-    // fields that require special handling
-    if (key === 'alias' && (rootPath === 'resolve' || rootPath === '')) {
-      merged[key] = mergeAlias(existing, value)
-      continue
-    } else if (key === 'assetsInclude' && rootPath === '') {
-      merged[key] = [].concat(existing, value)
-      continue
-    } else if (
-      key === 'noExternal' &&
-      rootPath === 'ssr' &&
-      (existing === true || value === true)
-    ) {
-      merged[key] = true
-      continue
-    }
-
-    if (Array.isArray(existing) || Array.isArray(value)) {
-      merged[key] = [...arraify(existing ?? []), ...arraify(value ?? [])]
-      continue
-    }
-    if (isObject(existing) && isObject(value)) {
-      merged[key] = mergeConfigRecursively(
-        existing,
-        value,
-        rootPath ? `${rootPath}.${key}` : key
-      )
-      continue
-    }
-
-    merged[key] = value
-  }
-  return merged
-}
-
-export function mergeConfig(
-  defaults: Record<string, any>,
-  overrides: Record<string, any>,
-  isRoot = true
-): Record<string, any> {
-  return mergeConfigRecursively(defaults, overrides, isRoot ? '' : '.')
-}
-
-function mergeAlias(
-  a?: AliasOptions,
-  b?: AliasOptions
-): AliasOptions | undefined {
-  if (!a) return b
-  if (!b) return a
-  if (isObject(a) && isObject(b)) {
-    return { ...a, ...b }
-  }
-  // the order is flipped because the alias is resolved from top-down,
-  // where the later should have higher priority
-  return [...normalizeAlias(b), ...normalizeAlias(a)]
-}
-
-function normalizeAlias(o: AliasOptions = []): Alias[] {
-  return Array.isArray(o)
-    ? o.map(normalizeSingleAlias)
-    : Object.keys(o).map((find) =>
-        normalizeSingleAlias({
-          find,
-          replacement: (o as any)[find]
-        })
-      )
-}
-
-// https://github.com/vitejs/vite/issues/1363
-// work around https://github.com/rollup/plugins/issues/759
-function normalizeSingleAlias({
-  find,
-  replacement,
-  customResolver
-}: Alias): Alias {
-  if (
-    typeof find === 'string' &&
-    find.endsWith('/') &&
-    replacement.endsWith('/')
-  ) {
-    find = find.slice(0, find.length - 1)
-    replacement = replacement.slice(0, replacement.length - 1)
-  }
-
-  const alias: Alias = {
-    find,
-    replacement
-  }
-  if (customResolver) {
-    alias.customResolver = customResolver
-  }
-  return alias
-}
 
 export function sortUserPlugins(
   plugins: (Plugin | Plugin[])[] | undefined
