+++ packages/vite/src/node/public-utils.ts
@@ -1,0 +1,13 @@
+/**
+ * Exported sync utils should go here.
+ * This file will be bundled to ESM and CJS and redirected by ../index.cjs
+ * Please control the side-effects by checking the ./dist/node-cjs/utils.cjs bundle
+ */
+export {
+  splitVendorChunkPlugin,
+  splitVendorChunk
+} from './plugins/splitVendorChunk'
+export { normalizePath, mergeConfig, mergeAlias } from './utils'
+export { send } from './server/send'
+export { createLogger } from './logger'
+export { searchForWorkspaceRoot } from './server/searchRoot'
