+++ packages/vite/src/node/server/index.ts
@@ -15,7 +15,8 @@ import type { SourceMap } from 'rollup'
 import type { CommonServerOptions } from '../http'
 import { httpServerStart, resolveHttpServer, resolveHttpsConfig } from '../http'
 import type { InlineConfig, ResolvedConfig } from '../config'
-import { mergeConfig, resolveConfig } from '../config'
+import { resolveConfig } from '../config'
+import { mergeConfig } from '../utils'
 import { isParentDirectory, normalizePath } from '../utils'
 import { ssrLoadModule } from '../ssr/ssrModuleLoader'
 import { resolveSSRExternal } from '../ssr/ssrExternal'
